requirements.txt
aiofiles==24.1.0
aioice==0.10.1
aiortc==1.14.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
argcomplete==2.0.0
asgiref==3.6.0
astroid==2.14.2
asttokens==2.2.1
attrs==22.2.0
audioread==3.1.0
av==16.0.1
Babel==2.10.3
beautifulsoup4==4.11.2
blinker==1.5
boto3==1.41.4
botocore==1.41.4
brotli==1.2.0
certifi==2022.9.24
cffi==2.0.0
chardet==5.1.0
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
colorzero==2.0
contextlib2==0.6.0.post1
contourpy==1.3.3
cryptography==46.0.3
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.3.2
decorator==5.2.1
deprecation==2.1.0
dill==0.3.6
distlib==0.3.6
distro==1.8.0
dnspython==2.8.0
docutils==0.19
fastapi==0.121.2
fastrtc==0.0.33
ffmpy==1.0.0
filelock==3.9.0
Flask==2.2.2
fonttools==4.60.1
fsspec==2025.10.0
future==0.18.2
gbinder-python==1.1.2
google-crc32c==1.7.1
gpiozero==2.0.1
gradio==5.49.1
gradio_client==1.13.3
groovy==0.1.2
h11==0.16.0
hailo-apps @ git+https://github.com/hailo-ai/hailo-apps-infra.git@2ecfc8662972ad8384546b7dbe79125dfddcb276
hailort==4.20.0
hf-xet==1.2.0
html5lib==1.1
httpcore==1.0.9
httplib2==0.20.4
httpx==0.28.1
huggingface_hub==1.1.4
idna==3.3
ifaddr==0.2.0
importlib-metadata==4.12.0
iniconfig==2.3.0
isort==5.6.4
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.5.2
jsonpointer==2.3
jsonschema==4.10.3
kiwisolver==1.4.9
lance-namespace==0.0.21
lance-namespace-urllib3-client==0.0.21
lancedb==0.25.3
lazr.restfulclient==0.14.5
lazr.uri==1.0.6
lazy-object-proxy==1.9.0
lazy_loader==0.4
lgpio==0.2.2.0
libarchive-c==2.9
libevdev==0.5
librosa==0.11.0
llvmlite==0.45.1
logilab-common==1.9.8
lxml==4.9.2
Mako==1.2.4.dev0
Markdown==3.4.1
markdown-it-py==4.0.0
MarkupSafe==2.1.2
matplotlib==3.10.7
mccabe==0.7.0
mdurl==0.1.2
meson==1.5.1
more-itertools==8.10.0
msgpack==1.1.2
mypy==1.0.1
mypy-extensions==0.4.3
netaddr==0.8.0
netifaces==0.11.0
numba==0.62.1
numpy==1.26.4
oauthlib==3.2.2
olefile==0.46
opencv-python==4.11.0.86
orjson==3.11.4
overrides==7.7.0
packaging==25.0
pandas==2.3.3
parso==0.8.3
pexpect==4.8.0
pgzero==1.2
picamera2==0.3.31
pidng==4.0.9
piexif==1.1.3
pigpio==1.78
Pillow==9.4.0
platformdirs==2.6.0
pluggy==1.6.0
pooch==1.8.2
psutil==5.9.4
ptyprocess==0.7.0
py==1.11.0
pyarrow==22.0.0
pycairo==1.20.1
pycparser==2.23
pycryptodomex==3.11.0
pycups==2.0.1
pydantic==2.11.10
pydantic_core==2.33.2
pydub==0.25.1
pyee==13.0.0
pygame==2.1.2
Pygments==2.14.0
PyGObject==3.42.2
pyinotify==0.9.6
PyJWT==2.6.0
pylance==0.39.0
pylibsrtp==1.0.0
pylint==2.16.2
PyOpenGL==3.1.6
pyOpenSSL==25.3.0
pyparsing==3.2.5
PyQt5==5.15.9
PyQt5-sip==12.11.1
pyrsistent==0.18.1
pyserial==3.5
pysmbc==1.0.23
pytest==9.0.1
pytest-mock==3.15.1
pytest-timeout==2.4.0
python-apt==2.6.0
python-dateutil==2.9.0.post0
python-dotenv==0.21.0
python-multipart==0.0.20
python-prctl==1.8.1
pytz==2022.7.1
pyudev==0.24.0
PyYAML==6.0
reportlab==3.6.12
requests==2.28.1
requests-oauthlib==1.3.0
responses==0.18.0
rfc3987==1.3.8
rich==14.2.0
roman==3.3
rpi-keyboard-config==1.0
rpi-lgpio==0.6
RTIMULib==7.2.1
ruff==0.14.5
s3transfer==0.15.0
safehttpx==0.1.7
scikit-learn==1.7.2
scipy==1.16.3
semantic-version==2.10.0
Send2Trash==1.8.1b0
sense-hat==2.6.0
setproctitle==1.3.7
shellingham==1.5.4
simplejpeg==1.8.1
simplejson==3.18.3
six==1.16.0
smbus2==0.4.2
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.3.2
soxr==1.0.0
spidev==3.5
ssh-import-id==5.10
starlette==0.49.3
tappas-core-python-binding @ file:///home/inha-vl/Downloads/yolo/hailo-rpi5-examples/hailo_temp_resources/tappas_core_python_binding-3.31.0-py3-none-any.whl#sha256=a4c8d0f26b2e1d6942c188114f0ae1ed0fc5bada54e3fc718d7bdfddc0ad8b4b
thonny==4.1.4
threadpoolctl==3.6.0
toml==0.10.2
tomlkit==0.13.3
tqdm==4.64.1
twython==3.8.2
typer==0.20.0
typer-slim==0.20.0
types-aiofiles==22.1
types-annoy==1.17
types-appdirs==1.4
types-aws-xray-sdk==2.10
types-babel==2.11
types-backports.ssl-match-hostname==3.7
types-beautifulsoup4==4.11
types-bleach==5.0
types-boto==2.49
types-braintree==4.17
types-cachetools==5.2
types-caldav==0.10
types-certifi==2021.10.8
types-cffi==1.15
types-chardet==5.0
types-chevron==0.14
types-click-spinner==0.1
types-colorama==0.4
types-commonmark==0.9
types-console-menu==0.7
types-contextvars==2.4
types-croniter==1.3
types-cryptography==3.3
types-D3DShot==0.1
types-dateparser==1.1
types-DateTimeRange==1.2
types-decorator==5.1
types-Deprecated==1.2
types-dj-database-url==1.0
types-docopt==0.6
types-docutils==0.19
types-editdistance==0.6
types-emoji==2.1
types-entrypoints==0.4
types-first==2.0
types-flake8-2020==1.7
types-flake8-bugbear==22.10.27
types-flake8-builtins==2.0
types-flake8-docstrings==1.6
types-flake8-plugin-utils==1.3
types-flake8-rst-docstrings==0.2
types-flake8-simplify==0.19
types-flake8-typing-imports==1.14
types-Flask-Cors==3.0
types-Flask-SQLAlchemy==2.5
types-fpdf2==2.5
types-gdb==12.1
types-google-cloud-ndb==1.11
types-hdbcli==2.14
types-html5lib==1.1
types-httplib2==0.21
types-humanfriendly==10.0
types-invoke==1.7
types-JACK-Client==0.5
types-jmespath==1.0
types-jsonschema==4.17
types-keyboard==0.13
types-ldap3==2.9
types-Markdown==3.4
types-mock==4.0
types-mypy-extensions==0.4
types-mysqlclient==2.1
types-oauthlib==3.2
types-openpyxl==3.0
types-opentracing==2.4
types-paho-mqtt==1.6
types-paramiko==2.11
types-parsimonious==0.10
types-passlib==1.7
types-passpy==1.0
types-peewee==3.15
types-pep8-naming==0.13
types-Pillow==9.3
types-playsound==1.3
types-polib==1.1
types-prettytable==3.4
types-protobuf==3.20
types-psutil==5.9
types-psycopg2==2.9
types-pyaudio==0.2
types-PyAutoGUI==0.9
types-pycurl==7.45
types-pyfarmhash==0.3
types-pyflakes==2.5
types-Pygments==2.13
types-pyinstaller==5.6
types-PyMySQL==1.0
types-pynput==1.7
types-pyOpenSSL==22.1
types-pyRFC3339==1.1
types-PyScreeze==0.1
types-pysftp==0.2
types-pytest-lazy-fixture==0.6
types-python-crontab==2.6
types-python-dateutil==2.8
types-python-gflags==3.1
types-python-jose==3.3
types-python-nmap==0.7
types-python-slugify==6.1
types-pytz==2022.6
types-pyvmomi==7.0
types-pywin32==304
types-PyYAML==6.0
types-redis==4.3
types-regex==2022.10.31
types-requests==2.28
types-retry==0.9
types-Send2Trash==1.8
types-setuptools==65.5
types-simplejson==3.17
types-singledispatch==3.7
types-six==1.16
types-slumber==0.7
types-SQLAlchemy==1.4.43
types-stdlib-list==0.8
types-stripe==3.5
types-tabulate==0.9
types-termcolor==1.1
types-toml==0.10
types-toposort==1.7
types-tqdm==4.64
types-tree-sitter==0.20
types-tree-sitter-languages==1.5
types-ttkthemes==3.2
types-typed-ast==1.5
types-tzlocal==4.2
types-ujson==5.5
types-urllib3==1.26
types-vobject==0.9
types-waitress==2.1
types-whatthepatch==1.0
types-xmltodict==0.13
types-xxhash==3.0
types-zxcvbn==4.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.1.1
urllib3==1.26.12
uvicorn==0.38.0
videodev2==0.0.4
virtualenv==20.17.1+ds
wadllib==1.3.6
webcolors==1.11.1
webencodings==0.5.1
websockets==15.0.1
Werkzeug==2.2.2
wrapt==1.14.1
zipp==1.0.0
